Refactor leader-key bindings
There are three major changes here:
1. Moving back to `:prefix` based declarations instead of the
`:keymaps` approach.  For some reason `:keymaps` doesn't seem to
handle the `:which-key` integration the way I was expecting.

2. Updating the specified `:states` to include `treemacs` for certain
keybindings, so that leader binding work in treemacs windows

3. Cleaning up alignment of key declarations, so that the file itself
is easier to read.

Along with those big-ticket items, there may be some specific bindings
that have been updated and are shuffled into the mix.