Add DWIM helper for S-return in org-mode
This is lifted and modified from Sacha Chua's emacs configuration to
handle inserting newlines from different contexts in org-mode
files (such as when inside a list item or heading).  There are
probably going to be further tweaks I want to make to this behavior
over time, but for now it's a good starting point (and better than
what I was getting with default behaviors).