Experiment with showing icons in tab-bar SVG images
Getting the SVG icon to show up works fine (at least for `material`
icons; `boxicons` at least seems to have issues), but getting the
rendered SVG to abide by the `:crop-left` option doesn't appear to
work.  I'm not sure if this is an issue with the padding argument
being passed, or if it's something deeper with `svg-lib` itself.