Co-locate "appearance"-related functions with packages
This mostly in-lines function definitions with their associated
packages rather than having them as a standalone file in my `lib`
directory.  The big exception is the `lib/tab-bar.el` code which is
substantiative enough to retain its own file, but is now loaded under
the `use-package` declaration for the `tab-bar` package, instead of
loading it in `init.el`.