Move `posframe`-related config to a standalone file
I'm still not sure how I feel about using posframe.  Aesthetically I like the
appearance of the "floating" window, but getting it to behave the way I want
across packages has been a challenge.  Extracting it to a separate file allows
me to keep experimenting with it in the future without cluttering the rest of my
configuration today.