Fix segfault when loading tab-bar configuration
(Hopefully)

This re-organizes some of my custom defined functions into `lib/tab-bar.el` to
work around what appears to be a load order issue that causes Emacs to crash
during initialization.  The segfault itself is a heisenbug, so I'm not 100% sure
this has fixed the issue, but initial testing indicates that this at least
improves the situation.